Feat/1965 double spends d (#2051)
* Added todo for checking block assembly

* Cleaned up conflicting transactions from block assembly

* Added processing of conflicting transactions back into subtree after reorg

* Added check for no conflicts in block102a on start

* merge from master

* Finished double spend handling for longer chains of txs

* WIP SQL utxo working, still need Aerospike and Memory to be finished

* Reduce logging

* Allow nil cancelFn

* Make initAerospike() the same signature as utils.SetupTestPostgresContainer

* Make readyCh really ready!

* Attempt at de-dupping test logic

* Better handling of channel logic

* Fix daemon tests

* Remove BlockValidation HTTP

* Moved tests into single function

* Fixed postgres SQL query, refactored tests

* Make sure Errorf() and Fatalf() print some output in TestLogger

* Finished Aerospike conflicting tx handling

* Implemented conflicting handling on subtreeprocessor Reset, added deserialize function in subtrees for conflicting only

* Disabled txMetaCache for tests to save 10+ seconds

* Fixed tracer in double spend tester

* Updated tracing in process conflicting

* Adding script to run test_full tagged tests in series.

* Compile each package before running tests

* Rename tags to sequentially

* Fix compilation errors in tests

---------

+++ test/stores/utxo/aerospike/test_helper.go
@@ -142,18 +142,20 @@ func cleanDB(t *testing.T, client *uaerospike.Client, key *aerospike.Key, txs ..
 
 	policy := util.GetAerospikeWritePolicy(tSettings, 0, aerospike.TTLDontExpire)
 
-	_, err := client.Delete(policy, key)
-	require.NoError(t, err)
+	if key != nil {
+		_, err := client.Delete(policy, key)
+		require.NoError(t, err)
+	}
 
 	if coinbaseKey != nil {
-		_, err = client.Delete(policy, coinbaseKey)
+		_, err := client.Delete(policy, coinbaseKey)
 		require.NoError(t, err)
 	}
 
 	if len(txs) > 0 {
 		for _, tx := range txs {
 			key, _ = aerospike.NewKey(aerospikeNamespace, aerospikeSet, tx.TxIDChainHash()[:])
-			_, err = client.Delete(policy, key)
+			_, err := client.Delete(policy, key)
 			require.NoError(t, err)
 		}
 	}
